refactor/all-structure according to https://jasonwatmore.com/post/2018/09/11/react-basic-http-authentication-tutorial-example
