+++ src/components/App/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import './App.css';
 
-import Home from './pages/Home';
+import Home from '../../pages/Home';
 
 
 
@@ -13,4 +13,4 @@ function App() {
   );
 }
 
-export default App;
+export { App };
