+++ src/components/Item/index.js
@@ -1,0 +1,1 @@
+export * from './Item';
