+++ src/components/NavBar/index.js
@@ -1,0 +1,1 @@
+export * from './NavBar';
