+++ src/components/Story/index.js
@@ -1,0 +1,1 @@
+export * from './Story';
